Getting rid of some warning from MSVC++

+++ Bcps/examples/Blis/BlisBranchStrategyPseudo.cpp
@@ -293,7 +293,7 @@ BlisBranchStrategyPseudo::createCandBranchObjects(int numPassesLeft)
     // branch up and down to initialize pseudo-cost.
     //--------------------------------------------------
     
-    numFirsts = firstObjects.size();
+    numFirsts = static_cast<int> (firstObjects.size());
     if (numFirsts > 0) {
 
         //--------------------------------------------------
